Add logger frames
